storage/localstore: GC quantiles refactoring
